reduce coupling to mixin
